refactor(reorganize): Add new directory structure and task runner system
This commit moves the existing code to the legacy/ folder, and adds the new directory structure and
task runner system for Evolution UI. What must be done next is to migrate the existing components
from the legacy/ folder to the new source/ directory, and modify the existing components to adhere
to their new anatomy. This will require renaming files and some other menial tasks.